requirements.txt
tensorflow-gpu
imageio
imgaug
json-tricks
matplotlib
nni
numpy
opencv-python
Pillow==6.1.0
scikit-image
scikit-learn
scipy
simplejson
tensorflow-gpu
tqdm
